+++ community/src/main/java/efub/assignment/community/board/repository/BoardRepository.java
@@ -1,4 +1,4 @@
-package efub.assignment.community.board;
+package efub.assignment.community.board.repository;
 
 import efub.assignment.community.board.domain.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
