+++ community/src/main/java/efub/assignment/community/post/repository/PostRepository.java
@@ -1,4 +1,4 @@
-package efub.assignment.community.post;
+package efub.assignment.community.post.repository;
 
 import efub.assignment.community.post.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
